+++ src/app/ui/missions/BannerCampaign/BannerCarousel.style.ts
