+++ src/app/ui/missions/BannerCampaign/ResponsiveBannerSkeleton.tsx
